Split iOS progress screen sections
